requirements.txt
coverage==4.4.1
decorator==4.1.2
Django==1.11.5
djangorestframework==3.6.4
Faker==0.7.3
flake8==3.4.1
greenlet==0.4.12
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.11.0
mccabe==0.6.1
mixer==5.6.6
msgpack-python==0.4.8
neovim==0.1.13
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pytest==3.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
